requirements.txt
Flask>=2.2
Werkzeug>=2.2
psycopg2-binary>=2.9
gunicorn
qrcode
Pillow
opencv-python
numpy
python-dotenv
